+++ doc/changelog.rst
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+* :feature:`314` Added :func:`Session.get_total_num_tests <slash.core.session.Session.get_total_num_tests>` for returning the number of tests expected to run in a session
 * :feature:`312` Add before_session_start hook
 * :feature:`311` Support plugin methods avoiding hook registrations with ``registers_on(None)``
 * :feature:`308` Support registering private methods in plugins using ``registers_on``